refactor: map 페이지와 레이아웃 분리 및 정리
